Move config file outside of server directory.
Configuration are app-wide, not just server scoped.
