+++ server/lib/util.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-var cfg = require('../config.json');
+var cfg = require('../../config.json');
 
 exports.validNick = function(nickname) {
     var regex = /^\w*$/;
